requirements.txt
Pillow>=9.0.0
PyQt6==6.5.2
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.2
PyInstaller==5.13.0



